+++ Assets/Scripts/DifferentClasses/CAVEntities/Interactive/Chest/ChestController.cs
@@ -45,6 +45,6 @@ public class ChestEditor : InterObjEditor
         stats = obj.GetStats();
         EditorGUILayout.Space();
         EditorGUILayout.LabelField("Parametres");
-        EditorGUILayout.IntField("direction", stats.direction);
+        EditorGUILayout.IntField("direction", (int)stats.direction);
     }
 }
